Added files
Priority Queue is exactly what it sounds like, a priority Queue, a nice data structure that allows you to add weights to the data stored. In this case data with a smaller value is given a higher priority as that is how the A* algorithm works.

HelperMethods is actually just one method in this case, it psuedo-randomly shuffles an array with a given seed (so if you want to get the same randomization a 2nd time you can do so by inputting the same seed). This file could have been added to the MapGenerator file, maybe in a future commit ;)

MapGenerator is the brains and brawn of the program, it controls all the animation speed and options, map and maze generation options, and does the bulk of the work for the actual A* algorithm.